changed IDs to unsigned big integers

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('post_id')->nullable();
+            $table->unsignedBigInteger('post_id')->nullable();
             $table->foreign('post_id')->references('id')->on('posts');
             $table->string('name');
             $table->string('email')->unique();
